% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi-funcs.R
\name{bdi_compute}
\alias{bdi_compute}
\alias{bdi_compute_sum}
\alias{bdi_factorise}
\title{Calculate BDI scores}
\usage{
bdi_compute(
  data,
  cols = matches("bdi_[0-9][0-9]$"),
  max_missing = 0,
  prefix = "bdi_",
  keep_all = TRUE
)

bdi_compute_sum(data, cols = matches("bdi_[0-9][0-9]$"), max_missing = 0)

bdi_factorise(bdi_sum)
}
\arguments{
\item{data}{Data containing BDI data}

\item{cols}{Columns that contain BDI data}

\item{max_missing}{Maximum number of components allowed to be missing.
Defaults to "0", and will return \code{NA} if missing any question. If set to
\code{NULL} any missing component counts as 0, meaning if all BDI
components are missing, the sum is still 0, not \code{NA}.}

\item{prefix}{string to prefix column names of computed values}

\item{keep_all}{logical, append to data.frame}

\item{bdi_sum}{Sum of BDI questions, as summed by \link{bdi_compute_sum}}
}
\value{
data.frame
}
\description{
{
Beck Depression Inventory-II (BDI-II) is one of the most widely used instruments for measuring the severity of self-reported depression in adolescents and adults.
As a general rule, BDI-II is administrated in LCBC to adults with an upper cut off around 60 years, while depression in older adults is assessed with the Geriatric Depression Scale (GDS).
However, please consult the instructions for each project, as this guideline has been implemented at different time points across the projects.

The questionnaire consists of 21 statements, each reflecting a depression symptom or attitude which could be rated from 0 to 3 in terms of intensity.
The answers should be based the participant´s feelings throughout the last week, including the day of filling out the form.
The sum of the scores of the items (0-3) yields one total score, with a possible range between 0 and 62.
\subsection{Interpretation of the scores (Total scores)}{\tabular{ll}{
   Score \tab Category \cr
   0-10 \tab These ups and downs are considered to be normal \cr
   11-16 \tab Mild mood disturbance \cr
   17 – 20 \tab Borderline clinical disturbance \cr
   21 – 30 \tab Moderate depression \cr
   31 – 40 \tab Severe depression \cr
   Above 40 \tab Extreme depression \cr
}


If a participant scores \verb{>= 17}, we should consider contacting the participant to follow up on this and offer making a note for the participant's doctor describing the scores.
}

\subsection{Data requirements}{
\subsection{Column names}{

By default, the functions assume that columns have names in the manner of \code{bdi_XX} where \code{XX} is a zero-padded (i.e. zero in front of numbers below 9, eg. \code{09}) question number of the inventory.
You may have column names in another format, but in that case you will need to supply to the functions the names of those columns using tidy-selectors (see the \href{https://www.tidyverse.org/}{tidyverse} packages for this).
The columns should adhere to some naming logic that is easy to specify.
}

\subsection{Data values}{

The values in the columns should be the item number of the question that was answered (i.e. \code{0}, \code{1}, \code{2}, or \code{3}).
The inventory allows subjects to respond to several options per question, in the case of this, the mean of the responded alternatives should be applied.
}

}

\subsection{References}{

Aaron T.Beck, Robert A.Steer, Margery G.Carbin (1988) Psychometric properties of the Beck Depression Inventory: Twenty-five years of evaluation, \emph{Clinical Psychology Review, Volume} 8, Issue 1, Pages 77-100, \href{https://doi.org/10.1016/0272-7358(88)90050-5}{doi: 10.1016/0272-7358(88)90050-5}

Robert A.Steer, David J.Rissmiller, Aaron T.Beck (2000) Use of the Beck Depression Inventory-II with depressed geriatric inpatients \emph{Behaviour Research and Therapy} Volume 38, Issue 3,Pages 311-318, \href{https://doi.org/10.1016/S0005-7967(99)00068-6}{doi: https://doi.org/10.1016/S0005-7967(99)00068-6}

Groth-Marnat G. (1990). \href{https://books.google.no/books?id=ldF5oRto-HAC&lpg=PR13&ots=Id0OKllubF&dq=Groth-Marnat\%20G.\%20(1990).\%20The\%20handbook\%20of\%20psychological\%20assessment\%20(2nd\%20ed.).\%20New\%20York\%3A\%20John\%20Wiley\%20\%26\%20Sons.&lr&hl=no&pg=PR13#v=onepage&q&f=false}{The handbook of psychological assessment (2nd ed.)}. New York: John Wiley & Sons.
}
}
}
\section{Functions}{
\itemize{
\item \code{bdi_compute_sum()}: Compute the BDI sum based on a data.frame containing the BDI data. Returns a numeric vector.

\item \code{bdi_factorise()}: Create a factor based on the BDI sum, with the cut-off points as described in original paper.

}}
\examples{
# Example of treatment of missing values
library(dplyr)
library(questionnaires)
data <- tibble(
bdi_01 = c(1, NA_real_, NA_real_, 2, 1),
bdi_02 = c(1, 1, NA_real_, 2, NA_real_)
)

# Row with all components missing, gets sum 0
  bind_cols(data,
      bdi_sum = bdi_compute_sum(data))

# Do not allow any missing values
  bind_cols(data,
      bdi_sum = bdi_compute_sum(data, max_missing = 0))
      
# Allow one missing value
  bind_cols(data,
      bdi_sum = bdi_compute_sum(data, max_missing = 2))
}
