% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_functions.R
\name{bdi_compute}
\alias{bdi_compute}
\title{Title}
\usage{
bdi_compute(data, cols = matches("BDI_[0-9][0-9]$"), keep_all = TRUE)
}
\arguments{
\item{data}{Data containing BDI data}

\item{cols}{Columns that contain BDI data}

\item{keep_all}{logical, append to data.frame}
}
\value{
data.frame
}
\description{
Title
}
\seealso{
Other bdi_functions: \code{\link{bdi_compute_sum}},
  \code{\link{bdi_factorise}}
}
\concept{bdi_functions}
