% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_functions.R
\name{bdi_factorise}
\alias{bdi_factorise}
\title{Factorize BDI sum}
\usage{
bdi_factorise(bdi_sum = sum)
}
\arguments{
\item{bdi_sum}{Sum of BDI questions, as summed by [bdi_compute_sum]}
}
\value{
factor
}
\description{
Factorize BDI sum
}
\seealso{
Other bdi_functions: 
\code{\link{bdi_compute_sum}()},
\code{\link{bdi_compute}()}
}
\concept{bdi_functions}
