% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_functions.R
\name{bdi_restructure}
\alias{bdi_restructure}
\title{Restructure BDI questions from wide format}
\usage{
bdi_restructure(data, cols = matches("[0-9]_[0-9]"), sep = "_")
}
\arguments{
\item{data}{Data containing BDI data}

\item{cols}{Columns that contain BDI data}

\item{sep}{separator to use for the column names}
}
\value{
data frame
}
\description{
If data come from Nettskjema, the structure is
in wide format, with each question option as
columns, creating 21*4 columns of data. 
This function allows you to gather and create
single columns for questions.
}
\details{
The columns must adhere to some specific logic to work.
It is recommended that the column names are in 
the format bdi_01_0 bdi_01_1 bdi_01_2 bdi_01_3,
where the first two numbers are the question 
number, and the last number is the option number.
}
\examples{
  dat <- data.frame(
     ID = 1:4, 
     bdi_01_0 = c(NA,1, NA, NA),
     bdi_01_1 = c(1, NA, 1, NA),
     bdi_01_2 = c(NA, NA, 1, NA),
     bdi_01_3 = c(NA, NA, NA, NA),
     bdi_02_0 = c(1, NA, NA, NA),
     bdi_02_1 = c(NA,NA, NA, NA),
     bdi_02_2 = c(NA,1, NA, NA),
     bdi_02_3 = c(NA, NA, NA, 1)
  )
  bdi_restructure(dat)
}
