% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfi2-funcs.R
\name{bfi_domain}
\alias{bfi_domain}
\alias{bfi_domain_extravers}
\alias{bfi_domain_agreeable}
\alias{bfi_domain_conscient}
\alias{bfi_domain_negemotion}
\alias{bfi_domain_openminded}
\title{BFI-2 Domain computations}
\usage{
bfi_domain_extravers(
  data,
  cols = matches("01$|06$|11$|16$|21$|26$|31$|36$|41$|46$|51$|56$"),
  reverse = TRUE,
  ...
)

bfi_domain_agreeable(
  data,
  cols = matches("02$|07$|12$|17$|22$|27$|32$|37$|42$|47$|52$|57$"),
  reverse = TRUE,
  ...
)

bfi_domain_conscient(
  data,
  cols = matches("03$|08$|13$|18$|23$|28$|33$|38$|43$|48$|53$|58$"),
  reverse = TRUE,
  ...
)

bfi_domain_negemotion(
  data,
  cols = matches("04$|09$|14$|19$|24$|29$|34$|39$|44$|49$|54$|59$"),
  reverse = TRUE,
  ...
)

bfi_domain_openminded(
  data,
  cols = matches("05$|10$|15$|20$|25$|30$|35$|40$|45$|50$|55$|60$"),
  reverse = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame with BFI data in}

\item{cols}{tidyselector(s) of which columns data are in}

\item{reverse}{logical. If reversal is needed (default) or not}

\item{...}{other arguments to \code{\link{bfi_reversal}}}
}
\value{
a vector with computed domain values.
}
\description{
{
Calculate the domains of the BFI-2
}
}
\section{Functions}{
\itemize{
\item \code{bfi_domain_extravers()}: Calculate the extraversion domain.

\item \code{bfi_domain_agreeable()}: Calculate the agreeableness domain.

\item \code{bfi_domain_conscient()}: Calculate the conscientiousness domain.

\item \code{bfi_domain_negemotion()}: Calculate the negative emotionality domain.

\item \code{bfi_domain_openminded()}: Calculate the open-minded domain.

}}
\examples{
library(dplyr)
# Making some test data
test_data <- dplyr::tibble(
  id = rep(1:10, each = 60),
  name = rep(sprintf("bfi_\%02d", 1:60), 10),
  value = lapply(1:10, function(x){
    sample(1:5, size = 60, replace = TRUE)
  }) \%>\% unlist()
) \%>\% 
  tidyr::pivot_wider()
  
bfi_domain_extravers(test_data)
bfi_domain_conscient(test_data)
}
