% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfi2-funcs.R
\name{bfi_facet}
\alias{bfi_facet}
\alias{bfi_facet_sociability}
\alias{bfi_facet_assertive}
\alias{bfi_facet_energy}
\alias{bfi_facet_compassion}
\alias{bfi_facet_respectful}
\alias{bfi_facet_trust}
\alias{bfi_facet_organization}
\alias{bfi_facet_productive}
\alias{bfi_facet_responsibility}
\alias{bfi_facet_anxiety}
\alias{bfi_facet_depression}
\alias{bfi_facet_emovolatility}
\alias{bfi_facet_intcuriosity}
\alias{bfi_facet_aestheticsens}
\alias{bfi_facet_imagination}
\title{BFI-2 Facet computations}
\usage{
bfi_facet_sociability(
  data,
  cols = matches("01$|16$|31$|46$"),
  reverse = TRUE,
  ...
)

bfi_facet_assertive(
  data,
  cols = matches("06$|21$|36$|51$"),
  reverse = TRUE,
  ...
)

bfi_facet_energy(data, cols = matches("11$|26$|41$|56$"), reverse = TRUE, ...)

bfi_facet_compassion(
  data,
  cols = matches("02$|17$|32$|47$"),
  reverse = TRUE,
  ...
)

bfi_facet_respectful(
  data,
  cols = matches("07$|22$|37$|52$"),
  reverse = TRUE,
  ...
)

bfi_facet_trust(data, cols = matches("12$|27$|42$|57$"), reverse = TRUE, ...)

bfi_facet_organization(
  data,
  cols = matches("03$|18$|33$|48$"),
  reverse = TRUE,
  ...
)

bfi_facet_productive(
  data,
  cols = matches("08$|23$|38$|53$"),
  reverse = TRUE,
  ...
)

bfi_facet_responsibility(
  data,
  cols = matches("13$|28$|43$|58$"),
  reverse = TRUE,
  ...
)

bfi_facet_anxiety(data, cols = matches("04$|19$|34$|49$"), reverse = TRUE, ...)

bfi_facet_depression(
  data,
  cols = matches("09$|24$|39$|54$"),
  reverse = TRUE,
  ...
)

bfi_facet_emovolatility(
  data,
  cols = matches("14$|29$|44$|59$"),
  reverse = TRUE,
  ...
)

bfi_facet_intcuriosity(
  data,
  cols = matches("10$|25$|40$|55$"),
  reverse = TRUE,
  ...
)

bfi_facet_aestheticsens(
  data,
  cols = matches("05$|20$|35$|50$"),
  reverse = TRUE,
  ...
)

bfi_facet_imagination(
  data,
  cols = matches("15$|30$|45$|60$"),
  reverse = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame with BFI data in}

\item{cols}{tidyselector(s) of which columns data are in}

\item{reverse}{logical. If reversal is needed (default) or not}

\item{...}{other arguments to \code{\link{bfi_reversal}}}
}
\value{
a vector with computed domain values.
}
\description{
{
Calculate the facets of the BFI-2
}
}
\section{Functions}{
\itemize{
\item \code{bfi_facet_sociability}: Calculate the sociability facet.

\item \code{bfi_facet_assertive}: Calculate the assertive facet.

\item \code{bfi_facet_energy}: Calculate the energy level facet.

\item \code{bfi_facet_compassion}: Calculate the compassion facet.

\item \code{bfi_facet_respectful}: Calculate the respectfulness facet.

\item \code{bfi_facet_trust}: Calculate the trust facet.

\item \code{bfi_facet_organization}: Calculate the organization facet.

\item \code{bfi_facet_productive}: Calculate the productive facet.

\item \code{bfi_facet_responsibility}: Calculate the responsibility facet.

\item \code{bfi_facet_anxiety}: Calculate the anxiety facet.

\item \code{bfi_facet_depression}: Calculate the depression facet.

\item \code{bfi_facet_emovolatility}: Calculate the emotional volatiliity facet.

\item \code{bfi_facet_intcuriosity}: Calculate the intellectual curiosity facet.

\item \code{bfi_facet_aestheticsens}: Calculate the aesthetic sensibility facet.

\item \code{bfi_facet_imagination}: Calculate the creative imagination facet.
}}

\examples{
# Making some test data
test_data <- dplyr::tibble(
  id = rep(1:10, each = 60),
  name = rep(sprintf("bfi_\%02d", 1:60), 10),
  value = lapply(1:10, function(x){
    sample(1:5, size = 60, replace = TRUE)
  }) |> unlist()
) |> 
  tidyr::pivot_wider()
  
bfi_facet_sociability(test_data)
bfi_facet_assertive(test_data)
}
