% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfi2-funcs.R
\name{bfi_reversal}
\alias{bfi_reversal}
\title{Big-5 Item reversals}
\usage{
bfi_reversal(data, ...)
}
\arguments{
\item{data}{Data with big-5 columns}

\item{...}{Column selection to reverse}
}
\value{
data.frame with specified columns reversed
}
\description{
Big-5 Item reversals
}
\examples{
data <- dplyr::tibble(
 col_01 = c(1:5, 3, 5, 4),
 col_02 = c(1:5, 3, 5, 4)
)
bfi_reversal(data, col_01)
}
