% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{compute_comp3}
\alias{compute_comp3}
\title{Compute component PSQI 3}
\usage{
compute_comp3(hoursSleep)
}
\arguments{
\item{hoursSleep}{column name with hours of sleep (decimal hours) [PSQI_04]}
}
\description{
Compute component PSQI 3
}
