% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{compute_comp5}
\alias{compute_comp5}
\title{Compute component PSQI 5}
\usage{
compute_comp5(data, noSleep30min = PSQI_05a,
  sleepTroubles = matches("^PSQI_05[a-j]$"))
}
\arguments{
\item{data}{data frame with the data}

\item{noSleep30min}{column name with evaluation of sleep within 30min (0-3) [PSQI_05a]}

\item{sleepTroubles}{columns containing sleep problem evaluations (0-3) [PSQI_05[a-j] ]}
}
\description{
Compute component PSQI 5
}
