% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{compute_global}
\alias{compute_global}
\title{Compute Global PSQI component}
\usage{
compute_global(data, cols = matches("^PSQI_Comp[0-9]+_"))
}
\arguments{
\item{data}{Data frame containing PSQI components}

\item{cols}{columns containing the components}
}
\description{
Compute Global PSQI component
}
