% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{compute_time_in_bed}
\alias{compute_time_in_bed}
\title{Compute time in bed}
\usage{
compute_time_in_bed(risingtime, bedtime)
}
\arguments{
\item{risingtime}{column name with rising time (HH:MM:SS) [PSQI_03]}

\item{bedtime}{column name with bedtime (HH:MM:SS) [PSQI_01]}
}
\description{
Compute time in bed
}
