% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{edu10_factorize}
\alias{edu10_factorize}
\title{Factorize vector of 10 category education data.}
\usage{
edu10_factorize(x)
}
\arguments{
\item{x}{Vector of 10 category education data}
}
\description{
Factorize vector of 10 category education data.
}
\seealso{
Other education_functions: \code{\link{edu10_reduce}},
  \code{\link{edu10_to_years}},
  \code{\link{edu4_factorize}},
  \code{\link{edu4_to_years}}, \code{\link{edu_compiled}},
  \code{\link{education_compute}}
}
\concept{education_functions}
