% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{edu10_reduce}
\alias{edu10_reduce}
\title{Reduce 10 category education to 4}
\usage{
edu10_reduce(x)
}
\arguments{
\item{x}{Vector of 10 category education data}
}
\description{
Reduce 10 category education to 4
}
\seealso{
Other education_functions: \code{\link{edu10_factorize}},
  \code{\link{edu10_to_years}},
  \code{\link{edu4_factorize}},
  \code{\link{edu4_to_years}}, \code{\link{edu_compiled}},
  \code{\link{education_compute}}
}
\concept{education_functions}
