% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{edu10_to_years}
\alias{edu10_to_years}
\title{Alter factorized education in 10 categories to years}
\usage{
edu10_to_years(x)
}
\arguments{
\item{x}{Vector of 10 category education data}
}
\value{
integer vector
}
\description{
Alter factorized education in 10 categories to years
}
\seealso{
Other education_functions: \code{\link{edu10_factorize}},
  \code{\link{edu10_reduce}}, \code{\link{edu4_factorize}},
  \code{\link{edu4_to_years}}, \code{\link{edu_compiled}},
  \code{\link{education_compute}}
}
\concept{education_functions}
