% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{edu4_factorize}
\alias{edu4_factorize}
\title{Factorise 4 category education data}
\usage{
edu4_factorize(x)
}
\arguments{
\item{x}{Vector of 4 category education data}
}
\description{
Factorise 4 category education data
}
\seealso{
Other education_functions: \code{\link{edu10_factorize}},
  \code{\link{edu10_reduce}}, \code{\link{edu10_to_years}},
  \code{\link{edu4_to_years}}, \code{\link{edu_compiled}},
  \code{\link{education_compute}}
}
\concept{education_functions}
