% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{edu4_to_years}
\alias{edu4_to_years}
\title{Turn 4 category education data to years}
\usage{
edu4_to_years(x)
}
\arguments{
\item{x}{Vector of 4 category education data}
}
\value{
vector of integers
}
\description{
Turn 4 category education data to years
}
\seealso{
Other education_functions: \code{\link{edu10_factorize}},
  \code{\link{edu10_reduce}}, \code{\link{edu10_to_years}},
  \code{\link{edu4_factorize}}, \code{\link{edu_compiled}},
  \code{\link{education_compute}}
}
\concept{education_functions}
