% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu-funcs.R
\name{edu_compile}
\alias{edu_compile}
\title{Compile education across sources}
\usage{
edu_compile(data, participant, mother, father)
}
\arguments{
\item{data}{MOAS-like data.frame}

\item{participant}{unquoted column of 4 category education for participant}

\item{mother}{unquoted column of 4 category education for participant's mother}

\item{father}{unquoted column of 4 category education for participant's father}
}
\value{
dataframe with three new columns
}
\description{
Compiles education from participant, mother or father depending
on source availability. Made for ease of testing and reporting education SES
of family
}
\examples{

edu <- data.frame(
    edu4 = c("3", "High school", 1, NA,
         "University/University college (> 4 years)", NA, 
          "University/University college (< 4 years)"),
    edu9 = c(7,7,8,NA,"Primary school (6 years)",5, 9),
    edu_years = c(NA, 12, 9, NA, 19, 19, NA),
    mother = c("3", "High school", 1, NA,
               "University/University college (> 4 years)",
               "University/University college (> 4 years)", 
               "University/University college (< 4 years)"),
    father = c(7,7,8,4,"Primary school (6 years)",5, 10),
    stringsAsFactors = FALSE
    )
 
 library(dplyr)
 edu |> 
     mutate(
        mother = ifelse(mother == "3", NA, mother),
        mother = edu4_factorise(mother),
        father = edu9_reduce(edu9_factorise(father))
     ) |> 
     edu_compile(
        participant = edu4,
        mother = mother,
        father = father
        )
}
\seealso{
Other edu_functions: 
\code{\link{edu_compute}()},
\code{\link{edu_factorise}()},
\code{\link{edu_levels2name}()},
\code{\link{edu_levels}()},
\code{\link{edu_reduce}()},
\code{\link{edu_to_years}()}
}
\concept{edu_functions}
