% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{edu_compiled}
\alias{edu_compiled}
\title{Compile education across sources}
\usage{
edu_compiled(data, participant, mother, father)
}
\arguments{
\item{data}{MOAS-like data.frame}

\item{participant}{unquoted column of 4 category education for participant}

\item{mother}{unquoted column of 4 category education for participant's mother}

\item{father}{unquoted column of 4 category education for participant's father}
}
\value{
dataframe with three new columns
}
\description{
Compiles education from participant, mother or father depending
on source availability. Made for ease of testing and reporting education SES
of family
}
\seealso{
Other education_functions: \code{\link{edu10_factorize}},
  \code{\link{edu10_reduce}}, \code{\link{edu10_to_years}},
  \code{\link{edu4_factorize}},
  \code{\link{edu4_to_years}},
  \code{\link{education_compute}}
}
\concept{education_functions}
