% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu_functions.R
\name{edu_factorise}
\alias{edu_factorise}
\alias{edu4_factorise}
\alias{edu9_factorise}
\title{Create factor from education vector}
\usage{
edu_factorise(x, levels)

edu4_factorise(x)

edu9_factorise(x)
}
\arguments{
\item{x}{character vector}

\item{levels}{levels returned from the edu_levels() function}
}
\value{
factor
}
\description{
Will convert even a mixed character
vector (combining numbers and text)
of education levels 10 and 4 to a 
factor.
}
\details{
Specialized returns
\itemize{
 \item{edu_factorise - }{with option to choose number of levels}
 \item{edu4_factorise - }{directly transform vector coded in 4-level scheme}
 \item{edu9_factorise - }{directly transform vector coded in 9-levels scheme}
}
}
\examples{

edu9 <- c("7", "7", "8", NA, "Primary school (6 years)", "5", "9")
edu_factorise(edu9, 9)
edu9_factorise(edu9)
}
\seealso{
Other edu_functions: \code{\link{edu_compile}},
  \code{\link{edu_compute}}, \code{\link{edu_levels2name}},
  \code{\link{edu_levels}}, \code{\link{edu_reduce}},
  \code{\link{edu_to_years}}
}
\concept{edu_functions}
