% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu-funcs.R
\name{edu_levels}
\alias{edu_levels}
\alias{edu4_levels}
\alias{edu9_levels}
\title{Get education levels scheme}
\usage{
edu_levels(levels = 4)

edu4_levels()

edu9_levels()
}
\arguments{
\item{levels}{how many levels to return (either 4 or 9)}
}
\value{
named numeric vector
}
\description{
Keeping track of the different educational
coding schemes at LCBC can be tricky.
This formula contains the two current types
of coding schemas employed by LCBC.
}
\details{
Specialized returns
\itemize{
\item{edu_levels - }{returns named numeric vector for levels specified}
\item{edu4_levels - }{returns named numeric vector for 4-levels scheme}
\item{edu9_levels - }{returns named numeric vector for 9-levels scheme}
}
}
\examples{
edu_levels(4)
edu_levels(9)

edu4_levels()
edu9_levels()
}
\seealso{
Other edu_functions: 
\code{\link{edu_compile}()},
\code{\link{edu_compute}()},
\code{\link{edu_factorise}()},
\code{\link{edu_levels2name}()},
\code{\link{edu_reduce}()},
\code{\link{edu_to_years}()}
}
\concept{edu_functions}
