% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu_functions.R
\name{edu_levels2name}
\alias{edu_levels2name}
\alias{edu4_levels2name}
\alias{edu9_levels2name}
\title{Alter levels to name}
\usage{
edu_levels2name(x, levels)

edu4_levels2name(x)

edu9_levels2name(x)
}
\arguments{
\item{x}{vector containing levels}

\item{levels}{numeric of number of levels (4 or 10)}
}
\value{
character vector
}
\description{
Change educational coded levels
to names of the levels
}
\details{
Specialized returns
\itemize{
 \item{edu_levels2name - }{transforms levels to names for levels specified}
 \item{edu4_levels2name - }{transforms levels to names for 4-levels scheme}
 \item{edu9_levels2name - }{transforms levels to names for 9-levels scheme}
}
}
\examples{
edu4 <- c(9, 9, 16, 19)
edu_levels2name(edu4, 4)
# does the same as
edu4_levels2name(edu4)

edu9 <- c(0, 6, 21, 16)
edu_levels2name(edu9, 9)
# does the same as
edu9_levels2name(edu9)
}
\seealso{
Other edu_functions: 
\code{\link{edu_compile}()},
\code{\link{edu_compute}()},
\code{\link{edu_factorise}()},
\code{\link{edu_levels}()},
\code{\link{edu_reduce}()},
\code{\link{edu_to_years}()}
}
\concept{edu_functions}
