% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu_functions.R
\name{edu_recode}
\alias{edu_recode}
\title{Recode new 9 levels into old}
\usage{
edu_recode(x, names = TRUE)
}
\arguments{
\item{x}{character vector of old scheme}

\item{names}{logical. toggle return of names rather than numbers}
}
\value{
character
}
\description{
New nettskjema data requires codebook to
not have special characters, and as such the old and
new coding scheme does not fit. This function turns new
coding scheme into the old, wanted one
}
\examples{
eds <- c(NA, "UnderGrad_BA", "HighSchool_Initial", "PostGrad_MA", 
   "PostGrad_PhD", "HighSchool", "Junior-HighSchool", "HighSchool_addition")
edu_recode(eds) 

eds <- c(1,5,8,2,6,9,1,10)
edu_recode(eds, names = FALSE) 
}
