% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu_functions.R
\name{edu_reduce}
\alias{edu_reduce}
\alias{edu9_reduce}
\title{Reduce education categories}
\usage{
edu_reduce(x, from, to)

edu9_reduce(x, to = 4)
}
\arguments{
\item{x}{character vector}

\item{from}{factor level to transform from}

\item{to}{factor level to transform to}
}
\value{
factor
}
\description{
These functions will aid in converting
one education scheme into another.
While you may attempt to go from a low
level to a high (from 4 to 9), there is
no way to actually do that in a consistent
way that will correctly reflect the
underlying data.
}
\details{
Always go from a higher level scheme to
a lower one (currently from 9 to 4 only)

Specialized returns
\itemize{
\item{edu_reduce - }{reduce with own to and from specification}
\item{edu9_reduce - }{directly reduce from 9 to 4}
}
}
\examples{
edu9 <- c("7", "7", "8", NA, "Primary school (6 years)", "5", "9")
edu_reduce(edu9, 9, 4)
edu9_reduce(edu9)
}
\seealso{
Other edu_functions: 
\code{\link{edu_compile}()},
\code{\link{edu_compute}()},
\code{\link{edu_factorise}()},
\code{\link{edu_levels2name}()},
\code{\link{edu_levels}()},
\code{\link{edu_to_years}()}
}
\concept{edu_functions}
