% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu_functions.R
\name{edu_to_years}
\alias{edu_to_years}
\alias{edu4_to_years}
\alias{edu9_to_years}
\title{Turn education data to years}
\usage{
edu_to_years(x, levels)

edu4_to_years(x)

edu9_to_years(x)
}
\arguments{
\item{x}{character vector}

\item{levels}{levels returned from the edu_levels() function}
}
\value{
vector of integers
}
\description{
Turn education data to years
}
\details{
Specialized returns
\itemize{
 \item{edu_to_years - }{Alter education to years specifying number of levels}
 \item{edu4_to_years - }{directly alter 4-level coded education to years }
 \item{edu9_to_years - }{directly alter 9-level coded education to years }
}
}
\examples{

edu4 <- c("3", "High school", "1", NA, 
          "University/University college (> 4 years)", 
           NA, "University/University college (< 4 years)")

edu_to_years(edu4, 4)
edu4_to_years(edu4)

edu9 <- c("7", "7", "8", NA, "Primary school (6 years)", "5", "9")
edu_to_years(edu9, 9)
edu9_to_years(edu9)
}
\seealso{
Other edu_functions: 
\code{\link{edu_compile}()},
\code{\link{edu_compute}()},
\code{\link{edu_factorise}()},
\code{\link{edu_levels2name}()},
\code{\link{edu_levels}()},
\code{\link{edu_reduce}()}
}
\concept{edu_functions}
