% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_functions.R
\name{education_compute}
\alias{education_compute}
\title{Fill inn Education in MOAS}
\usage{
education_compute(data, edu4 = Edu_Coded4, edu10 = Edu_Coded10,
  edu_years = Edu_Years, keep_all = TRUE)
}
\arguments{
\item{data}{MOAS-like data}

\item{edu4}{unquoted column containing Education coded in 4 categories}

\item{edu10}{unquoted column containing Education coded in 4 categories}

\item{edu_years}{unquoted column containing Education in years to highest completed}

\item{keep_all}{logical, append to data.frame}
}
\value{
a data.frame
}
\description{
Using existing data in the MOAS, fills in gaps, 
converts from on type of coding to another etc.
}
\seealso{
Other education_functions: \code{\link{edu10_factorize}},
  \code{\link{edu10_reduce}}, \code{\link{edu10_to_years}},
  \code{\link{edu4_factorize}},
  \code{\link{edu4_to_years}}, \code{\link{edu_compile}}
}
\concept{education_functions}
