% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi-funcs.R
\name{ehi_change}
\alias{ehi_change}
\title{Create vector with only correct values}
\usage{
ehi_change(x, direction = 1)
}
\arguments{
\item{x}{numeric vector}

\item{direction}{either 1 for positive, -1 for negative}
}
\value{
numeric vector
}
\description{
Since the coding we have often uses
negative numbers to indicate left-hand
preferences, a specialized function is here
to return a vector with only the values asked
for.
}
\details{
If direction is set to 1, returns only positive
numbers, negative and 0 returns as NA.
If direction is set to -1, returns only negative
numbers, positive and 0 returns as NA.
}
