% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_compute}
\alias{ehi_compute}
\title{Edinburgh handedness inventory}
\usage{
ehi_compute(data, cols = matches("^EHI_[0-9][0-9]$"), writing = EHI_01,
  keep_all = TRUE)
}
\arguments{
\item{data}{}

\item{writing}{}

\item{keep_all}{}

\item{positive_cols}{}

\item{negative_cols}{}
}
\value{
data.frame
}
\description{
Compute all variables of EHI
}
\seealso{
Other ehi_functions: \code{\link{ehi_compute_le}},
  \code{\link{ehi_factorise_le}},
  \code{\link{ehi_factorise_nominal}}
}
\concept{ehi_functions}
