% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_compute_le}
\alias{ehi_compute_le}
\title{Edinburgh handedness inventory LE}
\usage{
ehi_compute_le(data, cols = matches("^EHI_[0-9][0-9]$"))
}
\arguments{
\item{data}{data.frame containing EHI data}

\item{cols}{tidyselected columns of all EHI data}
}
\value{
numeric
}
\description{
Edinburgh handedness inventory LE
}
\seealso{
Other ehi_functions: \code{\link{ehi_compute}},
  \code{\link{ehi_factorise_le}},
  \code{\link{ehi_factorise_nominal}}
}
\concept{ehi_functions}
