% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_factorise_le}
\alias{ehi_factorise_le}
\title{Edinburgh handedness inventory code}
\usage{
ehi_factorise_le(le = EHI_LE)
}
\arguments{
\item{le}{numeric vector calculated by [ehi_compute_le]}
}
\value{
factor
}
\description{
Edinburgh handedness inventory code
}
\seealso{
Other ehi_functions: \code{\link{ehi_compute_le}},
  \code{\link{ehi_compute}},
  \code{\link{ehi_factorise_nominal}}
}
\concept{ehi_functions}
