% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_factorise_lq}
\alias{ehi_factorise_lq}
\alias{ehi_factorise_lqa}
\title{Factorise laterality quotient}
\usage{
ehi_factorise_lq(lq = ehi_lq)

ehi_factorise_lqa(
  lq,
  min = -70,
  max = 70,
  levels = c("left", "ambidexter", "right")
)
}
\arguments{
\item{lq}{numeric vector calculated by [ehi_compute_lq]}

\item{min}{minimum value for ambidexter specification (default = -70)}

\item{max}{maximum value for ambidexter specification (default = 70)}
}
\value{
factor
}
\description{
While the laterality quotient is nice 
to use if your sample and variance is
large enough for analyses, in most
cases you will need to report the categories
of laterality your participants fall within.
This function takes the laterality quotient as
computed by [ehi_compute_lq] and creates a factor
using common specifications.
}
\details{
\itemize{
 \item{ehi_factorise_lq - }{returns original two-factor specification}
 \item{ehi_factorise_lqa - }{returns commonly used three-factor specification}
}
}
\examples{
LQ <- c(1, 40, 70, -20, 0, 100, -90)
ehi_factorise_lq(LQ)
ehi_factorise_lqa(LQ)
ehi_factorise_lqa(LQ, min = -40, max = 60)
}
\seealso{
Other ehi_functions: 
\code{\link{ehi_compute_lq}()},
\code{\link{ehi_compute}()},
\code{\link{ehi_factorise_nominal}()}
}
\concept{ehi_functions}
