% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_factorise_nominal}
\alias{ehi_factorise_nominal}
\title{Edinburgh handedness inventory nominal}
\usage{
ehi_factorise_nominal(writing = EHI_01)
}
\arguments{
\item{writing}{numeric vector of writing prefereance [-2,-1,0,1,2]}
}
\value{
factor
}
\description{
Edinburgh handedness inventory nominal
}
\seealso{
Other ehi_functions: \code{\link{ehi_compute_le}},
  \code{\link{ehi_compute}}, \code{\link{ehi_factorise_le}}
}
\concept{ehi_functions}
