% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_values}
\alias{ehi_values}
\title{Sum ehi columns}
\usage{
ehi_values(data, cols = matches("^ehi_[0-9][0-9]$"), direction = 1)
}
\arguments{
\item{data}{data.frame containing ehi data}

\item{cols}{tidy-selection of all ehi columns}

\item{direction}{sum positive or negatives (1 for
positive, -1 for negative)}
}
\value{
numeric vector
}
\description{
Calculate the sum on non-NA
values in all columns in the specified direction(
1 == sum all positives, -1 sum absolutes values of
negatives)
}
