% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gad7}
\alias{gad7}
\title{Generalized Anxiety Disorder 7-item scale (GAD-7; Spitzer, Kroenke, Williams, & Löwe, 2006)}
\format{
\code{gad7}: a data frame with 7 rows each representing one item and the following columns:
\itemize{
\item \code{item_number}: Number of item
\item \code{measure}: Abbreviation of measure
\item \code{item_label_short}: Long form of item label
\item \code{item_label_long}: Sort form of item label
\item \code{item_value_labels}: List of labels for each item
}
}
\usage{
gad7
}
\description{
Generalized Anxiety Disorder 7-item scale (GAD-7; Spitzer, Kroenke, Williams, & Löwe, 2006)
}
\details{
Cronbach's Alpha (Internal Consistency): .92 (Spitzer, Kroenke, Williams, & Löwe, 2006)

Test-restest Reliability (1 week, n=236): .83 (Spitzer, Kroenke, Williams, & Löwe, 2006)

Reliable Change Index: 4 (IAPT Manual, 2020)

Labels are scored from 0 to 3 for all items.

\tabular{ll}{
Total Score \tab Anxiety Severity\cr
1-4 \tab Minimal anxiety \cr
5-9 \tab Mild anxiety \cr
10-14 \tab Moderate anxiety \cr
15-27 \tab Severe anxiety
}
}
\references{
Spitzer, R. L., Kroenke, K., Williams, J. B. W., & Löwe, B. (2006). A brief measure for assessing generalized anxiety disorder: The GAD-7. Archives of Internal Medicine, 166, 1092–1097. \url{https://doi.org/10.1001/archinte.166.10.1092}

The Improving Access to Psychological Therapies Manual (March 2020). Version number: 4 \url{https://www.england.nhs.uk/publication/the-improving-access-to-psychological-therapies-manual/}
}
\keyword{datasets}
