% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_functions.R
\name{gds_alter_values}
\alias{gds_alter_values}
\title{Change coding of GDS to correct numeric values}
\usage{
gds_alter_values(
  data,
  values = gds_values(),
  reverse = FALSE,
  cols = dplyr::matches("01$|05$|07$|09$|15$|19$|21$|27$|29$|30$")
)
}
\arguments{
\item{data}{data.frame with GDS data in it}

\item{values}{named vector of 2 providing the coding for Yes and No answers c(Yes = 1, No = 2)}

\item{reverse}{reverse logic}

\item{cols}{GDS data columns}
}
\description{
Necessary step for computing the total score
}
\seealso{
Other gds_functions: 
\code{\link{gds_binary}()},
\code{\link{gds_compute_sum}()},
\code{\link{gds_compute}()},
\code{\link{gds_factorise}()},
\code{\link{gds_values}()}
}
\concept{gds_functions}
