% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_functions.R
\name{gds_binary}
\alias{gds_binary}
\title{Binerise GDS values}
\usage{
gds_binary(x, values = gds_values())
}
\arguments{
\item{x}{vector of yes and no coding}

\item{values}{named vector of 2 providing the coding for Yes and No answers c(Yes = 1, No = 2)}
}
\value{
vector of 0's and 1's
}
\description{
internal function to make all "yes" answers
equal to 1, and all "no" to 0. This for
convenience of calculations later.
}
\examples{
gds_binary(c(1,1,0,NA,1), gds_values(1,0))
gds_binary(c("y","y","n",NA,"y"), gds_values(yes = "y", no = "n"))
}
\seealso{
Other gds_functions: 
\code{\link{gds_alter_values}()},
\code{\link{gds_compute_sum}()},
\code{\link{gds_compute}()},
\code{\link{gds_factorise}()},
\code{\link{gds_values}()}
}
\concept{gds_functions}
