% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_functions.R
\name{gds_factorise}
\alias{gds_factorise}
\title{Factorise the GDS sum}
\usage{
gds_factorise(gds_sum)
}
\arguments{
\item{gds_sum}{numeric vector of GDS sums}
}
\value{
factor
}
\description{
Factorise the GDS sum
}
\seealso{
Other gds_functions: 
\code{\link{gds_alter_values}()},
\code{\link{gds_binary}()},
\code{\link{gds_compute_sum}()},
\code{\link{gds_compute}()},
\code{\link{gds_values}()}
}
\concept{gds_functions}
