% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_functions.R
\name{gds_values}
\alias{gds_values}
\title{Specify coding scheme for GDS questions}
\usage{
gds_values(yes = 1, no = 0)
}
\arguments{
\item{yes}{value indicating a positive answer}

\item{no}{value indicating a negative answer}
}
\value{
list of yes and no values
}
\description{
Function to easily set the response coding used
in the GDS data.
}
\examples{
gds_values()
gds_values(yes = "YES", no = "NO")
}
