% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_functions.R
\name{income_nok2other}
\alias{income_nok2other}
\title{Translate NOK to other currency}
\usage{
income_nok2other(x, rate = 0.1)
}
\arguments{
\item{x}{currency}

\item{rate}{currency translation rate (defaul 0.10 for euro)}
}
\value{
numeric
}
\description{
In order to compare income in Norway
to other countries, currency concersions
might be necessary. This function multiplies
with the rate provided.
}
\examples{
income_nok2other(c(100, 2930, 13649))
income_nok2other(c(100, 2930, 13649), 0.5)
}
