% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq_functions.R
\name{ipaq_compute}
\alias{ipaq_compute}
\title{Compute all the IPAQ components}
\usage{
ipaq_compute(data, mets = list(light = 3.3, moderate = 4, vigorous = 8),
  light_days = IPAQ_1b, light_mins = IPAQ_2, mod_days = IPAQ_3b,
  mod_mins = IPAQ_4, vig_days = IPAQ_5b, vig_mins = IPAQ_6,
  keep_all = TRUE)
}
\arguments{
\item{data}{data.frame containing all the IPAQ data}

\item{mets}{list with three elements for the MET categories [default = list(light = 3.3, moderate = 4.0, vigorous = 8.0)]}

\item{light_days}{column with the days of light activity}

\item{light_mins}{column with the minutes of light activity}

\item{mod_days}{column with the days of moderate activity}

\item{mod_mins}{column with the minutes of moderate activity}

\item{vig_days}{column with the days of vigurous activity}

\item{vig_mins}{column with the minutes of vigurous activity}

\item{keep_all}{logical, append to data.frame}
}
\description{
Compute all the IPAQ components
}
