% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq_functions.R
\name{ipaq_compute_sum}
\alias{ipaq_compute_sum}
\title{Calculate the IPAQ sum}
\usage{
ipaq_compute_sum(vigurous, moderate, light)
}
\arguments{
\item{vigurous}{Vector with vigurous MET calculated}

\item{moderate}{Vector with moderate MET calculated}

\item{light}{Vector with light MET calculated}
}
\description{
Calculate the IPAQ sum
}
