% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq_functions.R
\name{ipaq_mets}
\alias{ipaq_mets}
\title{IPAQ mets}
\usage{
ipaq_mets(light = 3.3, moderate = 4, vigorous = 8)
}
\arguments{
\item{light}{numeric. default 3.3}

\item{moderate}{numeric. default 4.0}

\item{vigorous}{numeric. default 8.0}
}
\value{
list of three
}
\description{
IPAQ calculations require specification of 
met (resting metabolic rate), which are not
necessarily static values. While there are defaults
for each of the three categories, there should
be the possibility to alter these with newer research.
}
\details{
This is a convenience function if users need to
alter the default values for one or more of the categories
and is compatible with the remaining IPAQ functions
in this package.
}
\examples{
ipaq_mets()
ipaq_mets(moderate = 5.1)
}
\seealso{
Other ipaq_functions: 
\code{\link{ipaq_compute_met}()},
\code{\link{ipaq_compute_sum}()},
\code{\link{ipaq_compute}()},
\code{\link{ipaq_time_alter}()}
}
\concept{ipaq_functions}
