% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{is_hm}
\alias{is_hm}
\title{Utility function to locate hm columns}
\usage{
is_hm(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical vector of length==ncol(data)
}
\description{
\code{is_hm} locates columns that are time (hm) classes
}
\examples{
\dontrun{
is_hm(data)
}
}
