% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_comp2}
\alias{psqi_compute_comp2}
\title{Compute component PSQI 2}
\usage{
psqi_compute_comp2(minBeforeSleep, noSleep30min)
}
\arguments{
\item{minBeforeSleep}{column name with no. minutes before sleep (numeric) [PSQI_02]}

\item{noSleep30min}{column name with evaluation of sleep within 30min (0-3) [PSQI_05a]}
}
\description{
Compute component PSQI 2
}
