% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_comp5}
\alias{psqi_compute_comp5}
\title{compute component PSQI 5}
\usage{
psqi_compute_comp5(data, sleep_troubles = matches("^psqi_05[b-j]$"))
}
\arguments{
\item{data}{data frame with the data}

\item{sleep_troubles}{columns containing sleep problem evaluations (0-3) [psqi_05[b-j] ]}
}
\description{
compute component PSQI 5
}
\seealso{
Other psqi_functions: 
\code{\link{psqi_compute_comp2}()},
\code{\link{psqi_compute_comp3}()},
\code{\link{psqi_compute_comp4}()},
\code{\link{psqi_compute_comp7}()},
\code{\link{psqi_compute_global}()},
\code{\link{psqi_compute}()}
}
\concept{psqi_functions}
