% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{time_alter}
\alias{time_alter}
\title{Turn strings of H:M to time}
\usage{
time_alter(x, unit = "minute")
}
\arguments{
\item{x}{string of class lubridate::hms "HH:MM:SS"}

\item{unit}{unit to convert to}
}
\description{
Turn strings of H:M to time
}
\examples{
time <- c("02:33")
time_alter(time)
time_alter(time, "minute")
}
