% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-funcs.R
\name{time_deci2period}
\alias{time_deci2period}
\title{Turn time into period}
\usage{
time_deci2period(x, unit = "hour", type = "hm")
}
\arguments{
\item{x}{string of class lubridate::hms "HH:MM:SS"}

\item{unit}{unit to convert to}

\item{type}{hms or hm}
}
\value{
Period
}
\description{
Turn strings in class hms into
periods of time.
}
\examples{
time_deci2period(8.5)
time_deci2period(1.25, "minute")
}
