% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{time_factor}
\alias{time_factor}
\title{Factor time of day}
\usage{
time_factor(x)
}
\arguments{
\item{x}{character or hms vector}
}
\value{
factor vector
}
\description{
Takes a vector of HH:MM (HH:MM:SS) information and
categorises these by a 4 level factor of time of day.
}
\examples{
time_factor(c("12:23", "15:59", "22:10", "8:13"))
}
