% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-funcs.R
\name{time_hms2deci}
\alias{time_hms2deci}
\title{Turn string time into decimal}
\usage{
time_hms2deci(x, unit = "hour")
}
\arguments{
\item{x}{string of class lubridate::hms "HH:MM:SS"}

\item{unit}{unit to convert to}
}
\value{
numeric vector
}
\description{
Turn string time into decimal
}
\examples{
time <- lubridate::hms("02:33:12")
time_hms2deci(time)
time_hms2deci(time, "minute")
}
