% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{time_hms2deciH}
\alias{time_hms2deciH}
\title{Turn string time into decimal hour}
\usage{
time_hms2deciH(x, unit = "hour")
}
\arguments{
\item{x}{string of class lubridate::hms "HH:MM:SS"}

\item{unit}{unit to convert to}
}
\value{
numeric vector
}
\description{
Turn string time into decimal hour
}
\examples{
time <- lubridate::hms("02:33:12")
time_hms2deciH(time)
time_hms2deciH(time, "minute")
}
