% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{time_of_day}
\alias{time_of_day}
\title{Create list of time of day break points}
\usage{
time_of_day(morning = c(5, 12), afternoon = c(12, 17), evening = c(17, 21))
}
\arguments{
\item{morning}{vector of two for the hours where morning start or end in 24H}

\item{afternoon}{vector of two for the hours where afternoon start or end in 24H}

\item{evening}{vector of two for the hours where evening start or end in 24H}
}
\value{
list of fours times of day classifying the 24H of the day
}
\description{
Create list of time of day break points
}
\examples{
time_of_day()
}
