% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygocity-funcs.R
\name{zygo_calc}
\alias{zygo_calc}
\title{Zygocity - Calculate item}
\usage{
zygo_calc(x, type, n = "single", recode = TRUE)
}
\arguments{
\item{x}{integer vector of answers to one
of the questionnaire questions.
Should not be longer than 2.}

\item{type}{type of question the vector is
from. "drop", "stranger, "dexterity",
"voice", "eye", or "belief".}

\item{n}{string indicating number of twins in
the pair available. Either "single" or "pair".}

\item{recode}{logical indicating if data should be
recoded from 1-5(7) to -1. 0. 1.}
}
\value{
single value of calculated score based
on recoded vector and multiplied with
correct factor weight.
}
\description{
\subsection{Classification}{

This note contains a brief description of the algorithm used to determine zygocity in recruitment in the 2000s.\tabular{lll}{
   Name \tab Answer questions about... \tab Used for \cr
   Drop \tab You and your twin were like two drops of water in childhood \tab Pairs and singles \cr
   Stranger \tab Strangers had trouble telling the difference when you were children \tab Pairs and singles \cr
   Eye \tab Similarity in terms of eye color \tab Pairs \cr
   Voice \tab Similarity in terms of voice \tab Single \cr
   Dexter \tab Similarity in Dexterity \tab Pairs and Singles \cr
   Belief \tab What you believe yourself \tab Pairs and Singles \cr
}


"Single" twins here means those who have responded alone, i.e. there is no data available for both in the pair.
The similarity questions that are not found in the table above, e.g. whether or not family members had problems distinguishing the twins is not used in the classification.
}

\subsection{Weights}{

During calculations of the entire zygocity score, weights are applied to the different categories,
depending on whether one or both twins have responded to the questionnaire.\tabular{llrr}{
   Name \tab Answer questions about... \tab Factor single \tab Factor pair \cr
   Drop \tab You and your twin were like two drops of water \tab 1.494 \tab 2.111 \cr
   Stranger \tab Strangers had trouble seeing the difference \tab 0.647 \tab 0.691 \cr
   Eye \tab Similarity in terms of eye color \tab  \tab 0.394 \cr
   Voice \tab Similarity in terms of voice \tab 0.347 \tab  \cr
   Dexter \tab Dexterity Similarity \tab 0.458 \tab 0.366 \cr
   Belief \tab What you believe yourself \tab 0.417 \tab 0.481 \cr
    \tab Constant term in the formula \tab 0.007 \tab - 0.087 \cr
}

}
}
\examples{
zygo_calc(c(1), type = "eye")
zygo_calc(c(1,3), type = "belief", n = "pair")
zygo_calc(c(4), type = "voice")
}
