% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygocity-funcs.R
\name{zygo_recode}
\alias{zygo_recode}
\title{Zygocity - recode variables}
\usage{
zygo_recode(x, type)
}
\arguments{
\item{x}{vector of numbers, either 1:3 or 1:4}

\item{type}{Type of question to recode. Can either
be 05, 06, 07 or 08, or drop, stranger, dexterity,
voice, eye or belief.}
}
\value{
return a vector with 0, -1 or 1.
}
\description{
\subsection{Coding}{

"Form value" is the value the answer option has in the data file.
"Score value" is the value used in the algorithm when zygocity is calculated.\tabular{llrr}{
   Variable \tab Answer option \tab Form value \tab Score value \cr
   Drop \tab Like two drops of water \tab 1 \tab 1 \cr
    \tab Like most siblings \tab 2 \tab -1 \cr
    \tab Don't know \tab 3 \tab 0 \cr
   Stranger \tab Often \tab 1 \tab 1 \cr
    \tab Occasionally \tab 2 \tab 0 \cr
    \tab Never \tab 3 \tab -1 \cr
    \tab Don't know \tab 4 \tab 0 \cr
   Belief \tab Monozygotic \tab 1 \tab 1 \cr
    \tab Dizygotic \tab 2 \tab -1 \cr
    \tab Don't know \tab 3 \tab 0 \cr
   Eye, Voice & Dexter \tab Exactly the same \tab 1 \tab 1 \cr
    \tab Almost like \tab 2 \tab 0 \cr
    \tab Different \tab 3 \tab -1 \cr
    \tab Don't know \tab 4 \tab 0 \cr
}


No answer option is used directly in the calculations, only the score values.
In the following, it is these values (-1, 0 or 1) that are used in the algorithms.
E.g. has Drop in the formula value 1 for a positive answer to whether the twins were equal to two drops of water.
}
}
\examples{
zygo_recode(c(1:4, NA), type = "eye")
zygo_recode(c(1:4, NA), type = "voice")
zygo_recode(c(1:3, NA), type = "drop")
}
