% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygocity-funcs.R
\name{zygo_weighted}
\alias{zygo_weighted}
\title{Calculate weighted zygocity item}
\usage{
zygo_weighted(x, type, n = "single")
}
\arguments{
\item{x}{vector or recoded zygocity data (-1, 0, 1)}

\item{type}{string. one of 'drop', 'stranger',
'dexterity', 'voice', 'eye' or 'belief.'}

\item{n}{string, 'pair' if both twins have answered,
'single' if not.}
}
\value{
numeric vector of weighted data
}
\description{
Calculate the item score of a question.
Function takes a single vector, with information
on the question type and the twin type ('single'
or 'pair') and calculates the zygocity item score.
}
