% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_functions.R
\name{bdi_compute}
\alias{bdi_compute}
\title{Compute all BDI data from questionnaire}
\usage{
bdi_compute(
  data,
  cols = matches("bdi_[0-9][0-9]$"),
  max_missing = 0,
  prefix = "bdi_",
  keep_all = TRUE
)
}
\arguments{
\item{data}{Data containing BDI data}

\item{cols}{Columns that contain BDI data}

\item{max_missing}{Maximum number of components allowed to be missing.
Defaults to "0", and will return \code{NA} if missing any question. If set to
\code{NULL} any missing component counts as 0, meaning if all BDI
components are missing, the sum is still 0, not \code{NA}.}

\item{prefix}{string to prefix column names of computed values}

\item{keep_all}{logical, append to data.frame}
}
\value{
data.frame
}
\description{
Computes the sum and factorises the sum into 
the four BDI categories based on the sum
}
\seealso{
Other bdi_functions: 
\code{\link{bdi_compute_sum}()},
\code{\link{bdi_factorise}()}
}
\concept{bdi_functions}
