% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi_functions.R
\name{bdi_compute_sum}
\alias{bdi_compute_sum}
\title{Calculate sum of BDI}
\usage{
bdi_compute_sum(data, cols = matches("BDI_[0-9][0-9]$"),
  min.answered = 18)
}
\arguments{
\item{data}{Data containing BDI data}

\item{cols}{Columns that contain BDI data}

\item{min.answered}{Minimum number of questions answered to sum score}
}
\value{
numeric
}
\description{
Calculate sum of BDI
}
\seealso{
Other bdi_functions: \code{\link{bdi_compute}},
  \code{\link{bdi_factorise}}
}
\concept{bdi_functions}
