% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{compute_time_in_bed}
\alias{compute_time_in_bed}
\title{compute time in bed}
\usage{
compute_time_in_bed(
  risingtime,
  bedtime,
  risingtime_func = lubridate::hm,
  bedtime_func = lubridate::hm
)
}
\arguments{
\item{risingtime}{column name with rising time (HH:MM:SS) [psqi_03]}

\item{bedtime}{column name with bedtime (HH:MM:SS) [psqi_01]}

\item{risingtime_func}{function to convert time to \code{Period}}

\item{bedtime_func}{function to convert time to \code{Period}}
}
\description{
compute time in bed
}
