% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu-funcs.R
\name{edu_map}
\alias{edu_map}
\alias{edu_map_chr}
\alias{edu_map_num}
\title{Create a mapped table for conversion}
\usage{
edu_map(from = 9, to = 4)

edu_map_chr(from = 9, to = 4)

edu_map_num(from = 9, to = 4)
}
\arguments{
\item{from}{schema levels to convert from}

\item{to}{shcema levels to convtert to}
}
\description{
Converting from a high-level educational
coding to a lower level one is cumbersome.
This function bases it self in any coding
scheme specified in \code{\link{edu_levels}}
and tries
creating a conversion table between two
specified schemas.
}
\details{
Specialized returns
\itemize{
\item{edu_map - }{returns a data.frame of two named vectors}
\item{edu_map_chr - }{returns a data.frame with two character vectors}
\item{edu_map_num - }{returns a data.frame with two numeric vectors}
}
}
