% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_compute}
\alias{ehi_compute}
\title{Edinburgh handedness inventory}
\usage{
ehi_compute(data, cols = matches("^EHI_[0-9][0-9]$"), writing = EHI_01,
  keep_all = TRUE)
}
\arguments{
\item{data}{data.frame containing EHI data}

\item{cols}{tidyselected columns of all EHI data}

\item{writing}{numeric vector of writing prefereance [-2,-1,0,1,2]}

\item{keep_all}{logical, append to data.frame}
}
\value{
data.frame
}
\description{
Compute all variables of EHI
}
\seealso{
Other ehi_functions: \code{\link{ehi_compute_le}},
  \code{\link{ehi_factorise_le}},
  \code{\link{ehi_factorise_nominal}}
}
\concept{ehi_functions}
