% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_compute_lq}
\alias{ehi_compute_lq}
\title{Laterality Quotient}
\usage{
ehi_compute_lq(data, cols = matches("^EHI_[0-9][0-9]$"))
}
\arguments{
\item{data}{data.frame containing EHI data}

\item{cols}{tidyselected columns of all EHI data}
}
\value{
numeric
}
\description{
The laterality quotient is calculated
using all the answes on the EHI, with the
formula:
(pos-neg)/(pos+neg)*100 )
}
\seealso{
Other ehi_functions: \code{\link{ehi_compute}},
  \code{\link{ehi_factorise_lq}},
  \code{\link{ehi_factorise_nominal}}
}
\concept{ehi_functions}
