% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehi_functions.R
\name{ehi_factorise_nominal}
\alias{ehi_factorise_nominal}
\title{Nominal laterality factor}
\usage{
ehi_factorise_nominal(writing = ehi_01)
}
\arguments{
\item{writing}{numeric vector of writing preference (-2,-1,0,1,2)}
}
\value{
factor
}
\description{
Using the answers to the first question
on writing from the Edinburgh handedness inventory,
a nominal scale of three factors can be returned.
}
\examples{
writing <- c(2, 2, -1, 0, 1, -2)
ehi_factorise_nominal(writing)
}
\seealso{
Other ehi_functions: 
\code{\link{ehi_compute_lq}()},
\code{\link{ehi_compute}()},
\code{\link{ehi_factorise_lq}()}
}
\concept{ehi_functions}
