% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq_functions.R
\name{ipaq_compute}
\alias{ipaq_compute}
\title{Compute all the IPAQ components}
\usage{
ipaq_compute(
  data,
  mets = ipaq_mets(),
  light_days = ipaq_1b,
  light_mins = ipaq_2,
  mod_days = ipaq_3b,
  mod_mins = ipaq_4,
  vig_days = ipaq_5b,
  vig_mins = ipaq_6,
  prefix = "ipaq_",
  keep_all = TRUE
)
}
\arguments{
\item{data}{data.frame containing all the ipaq data}

\item{mets}{list generated with ipaq_mets() [default = ipaq_mets()]}

\item{light_days}{column with the days of light activity}

\item{light_mins}{column with the minutes of light activity}

\item{mod_days}{column with the days of moderate activity}

\item{mod_mins}{column with the minutes of moderate activity}

\item{vig_days}{column with the days of vigorous activity}

\item{vig_mins}{column with the minutes of vigorous activity}

\item{prefix}{string to prefix column names of computed values}

\item{keep_all}{logical, append to data.frame}
}
\value{
data.frame
}
\description{
In most cases you have all the IPAQ
data in a data frame and would like to
calculate and add all the data directly
to your data frame.
}
\details{
This function takes many arguments in order to
do that.
}
\seealso{
Other ipaq_functions: 
\code{\link{ipaq_compute_met}()},
\code{\link{ipaq_compute_sum}()},
\code{\link{ipaq_mets}()},
\code{\link{ipaq_time_alter}()}
}
\concept{ipaq_functions}
