% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq-funcs.R
\name{ipaq_compute_met}
\alias{ipaq_compute_met}
\alias{ipaq_compute_sum}
\alias{ipaq_compute}
\title{Compute met from IPAQ}
\usage{
ipaq_compute_met(minutes = ipaq_2, days = ipaq_1b, met = ipaq_mets()$light)

ipaq_compute_sum(vigorous, moderate, light)

ipaq_compute(
  data,
  mets = ipaq_mets(),
  light_days = ipaq_5b,
  light_mins = ipaq_6,
  mod_days = ipaq_3b,
  mod_mins = ipaq_4,
  vig_days = ipaq_1b,
  vig_mins = ipaq_2,
  prefix = "ipaq_",
  keep_all = TRUE
)
}
\arguments{
\item{minutes}{vector of numeric minutes}

\item{days}{vector of numeric days}

\item{met}{met number (light = 3.3, moderate = 4.0, vigorous = 8)}

\item{vigorous}{Vector with vigorous met calculated}

\item{moderate}{Vector with moderate met calculated}

\item{light}{Vector with light met calculated}

\item{data}{data.frame containing all the ipaq data}

\item{mets}{list generated with ipaq_mets() (default = ipaq_mets())}

\item{light_days}{column with the days of light activity}

\item{light_mins}{column with the minutes of light activity}

\item{mod_days}{column with the days of moderate activity}

\item{mod_mins}{column with the minutes of moderate activity}

\item{vig_days}{column with the days of vigorous activity}

\item{vig_mins}{column with the minutes of vigorous activity}

\item{prefix}{string to prefix column names of computed values}

\item{keep_all}{logical, append to data.frame}
}
\value{
data.frame
}
\description{
The purpose of the International Physical Activity questionnaires (IPAQ) is to provide a set of well-developed instruments that can be used internationally to obtain comparable estimates of physical activity.
There are two versions of the questionnaire.
The short version is suitable for use in national and regional surveillance systems and the long version provide more detailed information often required in research work or for evaluation purposes.
\subsection{Scoring}{

Scoring of the IPAQ is based on a metric called METs, which are multiples of the resting metabolic rate.
\href{https://docs.google.com/viewer?a=v&pid=sites&srcid=ZGVmYXVsdGRvbWFpbnx0aGVpcGFxfGd4OjE0NDgxMDk3NDU1YWRlZTM}{The IPAQ scoring description can be found here }
\subsection{Continuous Score}{

Expressed as MET-min per week: MET level x minutes of activity/day x days per week

MET levels:
\itemize{
\item \strong{Light} - 3.3 METs
\item \strong{Moderate} - 4.0 METs
\item \strong{Vigorous} - 8.0 METs
}

Total MET-minutes/week = Light (3.3 x min x days) + Mod (4.0 x min x days) + Vig (8.0 x min x days)
}

\subsection{Categorical Score}{

Three levels (categories) of physical activity are proposed:
\subsection{Category 1: Low}{

This is the lowest level of physical activity. Those individuals who not meet criteria for categories 2 or 3 are considered low/inactive.
}

\subsection{Category 2: Moderate}{

Any one of the following 3 criteria:
\itemize{
\item 3 or more days of vigorous activity of at least 20 minutes per day \strong{OR}
\item 5 or more days of moderate-intensity activity or walking of at least 30 minutes per day \strong{OR}
\item 5 or more days of any combination of walking, moderate-intensity or vigorous intensity activities achieving a minimum of at least 600 MET-min/week.
}
}

\subsection{Category 3: High}{

Any one of the following 2 criteria:
\itemize{
\item Vigorous-intensity activity on at least 3 days and accumulating at least 1500 MET-minutes/
week \strong{OR}
\item 7 or more days of any combination of walking, moderate-intensity or vigorous
intensity activities achieving a minimum of at least 3000 MET-minutes/week
}
}

}

}

\subsection{References}{

Depression Screening Scale: A Preliminary Report, \emph{J Psychiatr Res}, 17 (1), 37-49, \href{https://www.sciencedirect.com/science/article/abs/pii/0022395682900334?via\%3Dihub}{doi: 10.1016/0022-3956(82)90033-4}

E L Lesher 1, J S Berryhill (1994), Validation of the Geriatric Depression Scale -- Short Form Among Inpatients, \emph{J Clin Psychol}, 50 (2), 256-60,
\href{https://doi.org/10.1002/1097-4679(199403)50:2\%3C256::aid-jclp2270500218\%3E3.0.co;2-e}{doi: 10.1002/1097-4679(199403)50:2<256::aid-jclp2270500218>3.0.co;2-e}
}
}
\section{Functions}{
\itemize{
\item \code{ipaq_compute_met()}: Calculate mets of an activity type

\item \code{ipaq_compute_sum()}: Calculate the IPAQ sum based on activities and mets

}}
\examples{
ipaq_vig_mins <- c(60, 20, 60, 25, 90, 20, 0, 75, 60, 30)
ipaq_vig_days <- c(1, 3, 2, 5, 6, 1, 1, 2, 2, 4)
ipaq_compute_met(ipaq_vig_mins, ipaq_vig_days, met = 8.0)
light = c(1300, 300)
moderate = c(200, 400)
vigorous = c(0, 1300)
ipaq_compute_sum(vigorous , moderate, light)
}
\seealso{
Other ipaq_functions: 
\code{\link{ipaq_mets}()},
\code{\link{ipaq_time_alter}()}
}
\concept{ipaq_functions}
