% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq_functions.R
\name{ipaq_compute_met}
\alias{ipaq_compute_met}
\title{Compute MET from IPAQ}
\usage{
ipaq_compute_met(minutes = IPAQ_2, days = IPAQ_1b, met = 3.3)
}
\arguments{
\item{minutes}{vector of numeric minutes}

\item{days}{vector of numeric days}

\item{met}{MET number (light = 3.3, moderate = 4.0, vigurous = 8)}
}
\value{
numeric vector
}
\description{
Compute MET from IPAQ
}
