% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaq_functions.R
\name{ipaq_time_alter}
\alias{ipaq_time_alter}
\title{Alter the time instant columns to decminals}
\usage{
ipaq_time_alter(data, cols = c(IPAQ_2, IPAQ_4, IPAQ_6))
}
\arguments{
\item{data}{data with columns to alter}

\item{cols}{columns to alter}
}
\value{
data.frame
}
\description{
Alter the time instant columns to decminals
}
