% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi-funcs.R
\name{psqi_compute_comp2}
\alias{psqi_compute_comp2}
\alias{psqi_compute_comp3}
\alias{psqi_compute_comp4}
\alias{psqi_compute_comp5}
\alias{psqi_compute_comp7}
\alias{psqi_compute_global}
\alias{psqi_compute}
\title{Compute all PSQI components and global score}
\usage{
psqi_compute_comp2(min_before_sleep, no_sleep_30min)

psqi_compute_comp3(hours_sleep)

psqi_compute_comp4(hours_sleep, bedtime, risingtime, ...)

psqi_compute_comp5(data, sleep_troubles = matches("^psqi_05[b-j]$"))

psqi_compute_comp7(keep_awake, keep_enthused)

psqi_compute_global(data, cols = matches("comp[1-7]+_"), max_missing = 0)

psqi_compute(
  data,
  components = 1:7,
  bedtime = psqi_01,
  min_before_sleep = psqi_02,
  risingtime = psqi_03,
  hours_sleep = psqi_04,
  no_sleep_30min = psqi_05a,
  sleepquality = psqi_06,
  medication = psqi_07,
  keep_awake = psqi_08,
  keep_enthused = psqi_09,
  sleep_troubles = matches("^psqi_05[b-j]$"),
  max_missing = 0,
  ...,
  prefix = "psqi_",
  keep_all = TRUE
)
}
\arguments{
\item{min_before_sleep}{column name with no. minutes before sleep (numeric) (psqi_02)}

\item{no_sleep_30min}{column name with evaluation of sleep within 30min (0-3) (psqi_05a)}

\item{hours_sleep}{column name with hours of sleep (decimal hours) (psqi_04)}

\item{bedtime}{column name with bedtime (HH:MM:SS) (psqi_01)}

\item{risingtime}{column name with rising time (HH:MM:SS) (psqi_03)}

\item{...}{other arguments to \code{\link{psqi_compute_time_in_bed}}}

\item{data}{data frame}

\item{sleep_troubles}{columns containing sleep problem evaluations (0-3) (psqi_05(b-j) )}

\item{keep_awake}{column name with evaluation of staying awake (0-3) (psqi_08)}

\item{keep_enthused}{column name with evaluation of keeping enthusiastic (0-3) (psqi_09)}

\item{cols}{columns containing the components}

\item{max_missing}{Integer specifying the number of missing values to accept in the PSQI components,
before the global PSQI value is set to missing. Defaults to 0. If \code{max_missing > 0}, the
global PSQI value is computed by weighting each non-missing entry with \code{7 / (7 - max_missing)}.}

\item{components}{integer vector of components to calculate. If all 7, global is added also}

\item{sleepquality}{column name with evaluation of sleep quality (0-3) (psqi_06)}

\item{medication}{column name with use of sleep mediation (0-3) (psqi_07)}

\item{prefix}{string to prefix column names of computed values}

\item{keep_all}{logical, append to data.frame}
}
\value{
a data.frame containing only the calculated components
}
\description{
Despite the prevalence of sleep complaints among psychiatric patients, few questionnaires have been specifically designed to measure sleep quality in clinical populations.
The Pittsburgh Sleep Quality Index (PSQI) is a self-rated questionnaire which assesses sleep quality and disturbances over a 1-month time interval.
Nineteen individual items generate seven “component” scores: subjective sleep quality, sleep latency, sleep duration, habitual sleep efficiency, sleep disturbances, use of sleeping medication, and daytime dysfunction.
The sum of scores for these seven components yields one global score.
##Scoring\tabular{rll}{
   component \tab name \tab description \cr
   1 \tab subjective sleep quality \tab Answer to q 6 \cr
   2 \tab sleep latency \tab Scaled sum of number of minutes before sleep (q 2) and evaluation of sleep within 30min (q 5a), scaled to a 5 point scale \cr
   3 \tab sleep duration \tab Scaled score of number of hours before one falls asleep (q 4), scaled to a 5 point scale \cr
   4 \tab habitual sleep efficiency \tab hours of sleep (q 4) divided by bedtime (q 1) subtracted from rising time (q 3), and scaled to a 5 point scale \cr
   5 \tab sleep disturbances \tab Sum of evaluation of sleep within 30min (q 5a) and all remaining questions on sleep problems (q 5b-j), scaled to a 5 point scale \cr
   6 \tab use of sleeping medication \tab Answer to question on use of sleep medication (q 7) \cr
   7 \tab daytime dysfunction \tab Sum of evaluation of staying awake (q 8) and evaluation of keeping enthusiastic (q 9), scaled to a 5-point scale \cr
   global score \tab sum of the above. \tab If any of the above is not possible to calculate, the global sum is also not calculated \cr
}

\subsection{Data requirements}{
\subsection{Column names}{

Questions with multiple subquestions should be named in a similar manner, suffixed by the alphabetical index (\strong{psqi_5a}, \strong{psqi_5b} etc.).
For questions 5j and 10j, the frequency of occurence should have the names \strong{psqi_5j} and \strong{psqi_10e}, and the freehand explanations should have any type of \strong{suffix} after this to indicate a text answers (i.e. \strong{psqi_5j_Desc} or \strong{psqi_5j_string}, \strong{psqi_5j_freehand}).
As an example, LCBC has the following set-up:
\itemize{
\item \strong{psqi_1}
\item \strong{psqi_2}
\item \strong{psqi_3}
\item \strong{psqi_4}
\item \strong{psqi_5a} \strong{psqi_5b} \strong{psqi_5c} \strong{psqi_5d} \strong{psqi_5e} \strong{psqi_5f} \strong{psqi_5g} \strong{psqi_5h} \strong{psqi_5i} \strong{psqi_5j} \strong{psqi_5j_Coded} \strong{psqi_5j_Desc}
\item \strong{psqi_6}
\item \strong{psqi_7}
\item \strong{psqi_8}
\item \strong{psqi_9}
\item \strong{psqi_10} \strong{psqi_10a} \strong{psqi_10b} \strong{psqi_10c} \strong{psqi_10d} \strong{psqi_10e} \strong{psqi_10e_Desc} \strong{psqi_10e_Coded}
\item \strong{psqi_11a} \strong{psqi_11b} \strong{psqi_11c} \strong{psqi_11d}
}
}

\subsection{4-option questions coding}{

All 4-option questions need to be coded \strong{0-3}, not \strong{1-4}.
}

\subsection{Time formats}{

For question 1, 3 and 4 (bedtime, rising time, hours of sleep), data should be punched as "\strong{HH:MM}".
Question 2 should be punched as minutes in numbers.
}

}

\subsection{References}{

Buysse et al. (1989) \href{https://www.sciencedirect.com/science/article/pii/0165178189900474}{The Pittsburgh sleep quality index: A new instrument for psychiatric practice and research}, \emph{Psychiatry Research}, 28:2, 193-213
}
}
\section{Functions}{
\itemize{
\item \code{psqi_compute_comp2()}: calculate the component 2 (sleep latency)

\item \code{psqi_compute_comp3()}: calculate the component 3 (sleep duratione)

\item \code{psqi_compute_comp4()}: calculate the component 4 (habitual sleep efficiency)

\item \code{psqi_compute_comp5()}: calculate the component 5 (sleep disturbance)

\item \code{psqi_compute_comp7()}: calculate the component 7 (daytime dysfunction)

\item \code{psqi_compute_global()}: calculate the global scores, sum of all components

}}
\concept{psqi_functions}
