% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_comp2}
\alias{psqi_compute_comp2}
\title{Compute component PSQI 2}
\usage{
psqi_compute_comp2(minBeforeSleep, noSleep30min)
}
\arguments{
\item{minBeforeSleep}{column name with no. minutes before sleep (numeric) [PSQI_02]}

\item{noSleep30min}{column name with evaluation of sleep within 30min (0-3) [PSQI_05a]}
}
\description{
Computes the second PSQI component on sleep quality, 
using the number of minutes before sleep (PSQI Q2) and
subjective evaluation of sleep within 30 minutes
}
\seealso{
Other psqi_functions: 
\code{\link{psqi_compute_comp3}()},
\code{\link{psqi_compute_comp4}()},
\code{\link{psqi_compute_comp5}()},
\code{\link{psqi_compute_comp7}()},
\code{\link{psqi_compute_global}()},
\code{\link{psqi_compute}()}
}
\concept{psqi_functions}
