% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_comp3}
\alias{psqi_compute_comp3}
\title{Compute component PSQI 3}
\usage{
psqi_compute_comp3(hoursSleep)
}
\arguments{
\item{hoursSleep}{column name with hours of sleep (decimal hours) [PSQI_04]}
}
\description{
Compute component PSQI 3
}
\seealso{
Other psqi_functions: \code{\link{psqi_compute_comp2}},
  \code{\link{psqi_compute_comp4}},
  \code{\link{psqi_compute_comp5}},
  \code{\link{psqi_compute_comp7}},
  \code{\link{psqi_compute_global}},
  \code{\link{psqi_compute}}
}
\concept{psqi_functions}
