% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_comp4}
\alias{psqi_compute_comp4}
\title{Compute component PSQI 4}
\usage{
psqi_compute_comp4(hoursSleep, bedtime, risingtime)
}
\arguments{
\item{hoursSleep}{column name with hours of sleep (decimal hours) [PSQI_04]}

\item{bedtime}{column name with bedtime (HH:MM:SS) [PSQI_01]}

\item{risingtime}{column name with rising time (HH:MM:SS) [PSQI_03]}
}
\description{
Compute component PSQI 4
}
\seealso{
Other psqi_functions: \code{\link{psqi_compute_comp2}},
  \code{\link{psqi_compute_comp3}},
  \code{\link{psqi_compute_comp5}},
  \code{\link{psqi_compute_comp7}},
  \code{\link{psqi_compute_global}},
  \code{\link{psqi_compute}}
}
\concept{psqi_functions}
