% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_comp7}
\alias{psqi_compute_comp7}
\title{Calculate component 7 of PSQI}
\usage{
psqi_compute_comp7(keepAwake, keepEnthused)
}
\arguments{
\item{keepAwake}{column name with evaluation of staying awake (0-3) [PSQI_08]}

\item{keepEnthused}{column name with evaluation of keeping enthusiastic (0-3) [PSQI_09]}
}
\description{
Calculate component 7 of PSQI
}
\seealso{
Other psqi_functions: \code{\link{psqi_compute_comp2}},
  \code{\link{psqi_compute_comp3}},
  \code{\link{psqi_compute_comp4}},
  \code{\link{psqi_compute_comp5}},
  \code{\link{psqi_compute_global}},
  \code{\link{psqi_compute}}
}
\concept{psqi_functions}
