% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_global}
\alias{psqi_compute_global}
\title{Compute Global PSQI component}
\usage{
psqi_compute_global(data, cols = matches("^PSQI_Comp[1-7]+_"))
}
\arguments{
\item{data}{Data frame containing PSQI components}

\item{cols}{columns containing the components}
}
\description{
Compute Global PSQI component
}
\seealso{
Other psqi_functions: \code{\link{psqi_compute_comp2}},
  \code{\link{psqi_compute_comp3}},
  \code{\link{psqi_compute_comp4}},
  \code{\link{psqi_compute_comp5}},
  \code{\link{psqi_compute_comp7}},
  \code{\link{psqi_compute}}
}
\concept{psqi_functions}
