% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi_functions.R
\name{psqi_compute_global}
\alias{psqi_compute_global}
\title{Compute Global PSQI component}
\usage{
psqi_compute_global(data, cols = matches("^PSQI_Comp[1-7]+_"),
  max_missing = 0)
}
\arguments{
\item{data}{Data frame containing PSQI components}

\item{cols}{columns containing the components}

\item{max_missing}{Integer specifying the number of missing values to accept in the PSQI components, 
before the global PSQI value is set to missing. Defaults to 0. If \code{max_missing > 0}, the 
global PSQI value is computed by weighting each non-missing entry with \code{7 / (7 - max_missing)}.}
}
\description{
Compute Global PSQI component
}
\seealso{
Other psqi_functions: \code{\link{psqi_compute_comp2}},
  \code{\link{psqi_compute_comp3}},
  \code{\link{psqi_compute_comp4}},
  \code{\link{psqi_compute_comp5}},
  \code{\link{psqi_compute_comp7}},
  \code{\link{psqi_compute}}
}
\concept{psqi_functions}
