% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tas-funcs.R
\name{tas_compute}
\alias{tas_compute}
\title{Compute the TAS factors}
\usage{
tas_compute(
  data,
  reverse_cols = c(tas_04, tas_05, tas_10, tas_18),
  identify_cols = c(tas_01, tas_03, tas_06, tas_07, tas_09, tas_13, tas_14),
  describe_cols = c(tas_02, tas_04, tas_11, tas_12, tas_17),
  thinking_cols = c(tas_05, tas_08, tas_10, tas_15, tas_16, tas_18, tas_19, tas_20),
  prefix = "tas_",
  keep_all = TRUE
)
}
\arguments{
\item{data}{Data containing TAS data}

\item{reverse_cols}{Columns that need reversing}

\item{identify_cols}{Columns for the "identify feeling" factor}

\item{describe_cols}{Columns for the "describing feelings" factor}

\item{thinking_cols}{Columns for the "externally oriented thinking" factor}

\item{prefix}{string to prefix column names of computed values}

\item{keep_all}{logical, append to data.frame}
}
\description{
Compute the TAS factors
}
